// src/type/type.ts
type History = {
    id: string;
    name_es: string;
    name_en: string;
    name_pt: string;
    active: boolean;
    description_en: string;
    description_es: string;
    description_pt: string;
    youtube: string;
    poster?: string;
    cover?: string;
    type: number;
    isPremium: boolean;
    song?: string;
    createdAt?: string;
    updatedAt?: string;
};


export type { History } 